Move DialogBubble to index.js
